plugins: split ipc-rules into multiple files
Makes it easier to keep an overview of what is going on.
